feat: add tool to list calendar events within a specified date range
